+++ lib/valvat/lookup.rb
@@ -18,6 +18,7 @@ class Valvat
         response[:valid] && (options[:detail] || options[:requester_vat]) ? 
           filter_detail(response) : response[:valid]
       rescue => err
+        @last_error = err
         if err.respond_to?(:to_hash) && err.to_hash[:fault] && err.to_hash[:fault][:faultstring] == "{ 'INVALID_INPUT' }"
           return false
         end
@@ -38,6 +39,10 @@ class Valvat
           wsdl.document = 'http://ec.europa.eu/taxation_customs/vies/checkVatService.wsdl'
         end
       end
+    end
+    
+    def self.last_error
+      @last_error
     end
     
     private